+++ android/src/main/kotlin/fl/webview/FlWebChromeClient.kt
@@ -1,14 +1,20 @@
 package fl.webview
 
+import android.net.Uri
 import android.os.Handler
 import android.os.Looper
 import android.os.Message
+import android.util.Log
 import android.webkit.GeolocationPermissions
+import android.webkit.PermissionRequest
+import android.webkit.ValueCallback
 import android.webkit.WebChromeClient
 import android.webkit.WebResourceRequest
 import android.webkit.WebView
 import android.webkit.WebViewClient
 import io.flutter.plugin.common.MethodChannel
+import java.io.File
+import java.lang.Exception
 
 class FlWebChromeClient(
     private val channel: MethodChannel,
@@ -32,7 +38,6 @@ class FlWebChromeClient(
                 )
             }
         }
-
         val transport = resultMsg.obj as WebView.WebViewTransport
         transport.webView = newWebView
         resultMsg.sendToTarget()
@@ -46,18 +51,7 @@ class FlWebChromeClient(
         if (enabledProgressChanged) {
             if (lastProgress == progress || progress < lastProgress) return
             lastProgress = progress
-            invokeMethod("onProgress", progress)
-        }
-    }
-
-
-    private fun invokeMethod(method: String, args: Any?) {
-        if (handler.looper == Looper.myLooper()) {
-            channel.invokeMethod(method, args)
-        } else {
-            handler.post {
-                channel.invokeMethod(method, args)
-            }
+            FlWebViewPlugin.invokeMethod(channel, handler, "onProgress", progress)
         }
     }
 
@@ -66,6 +60,51 @@ class FlWebChromeClient(
         origin: String?, callback: GeolocationPermissions.Callback?
     ) {
         callback?.invoke(origin, true, false);
+    }
+
+
+    override fun onShowFileChooser(
+        webView: WebView?,
+        filePathCallback: ValueCallback<Array<Uri>>?,
+        fileChooserParams: FileChooserParams?
+    ): Boolean {
+        val params = mapOf(
+            "title" to fileChooserParams?.title,
+            "mode" to fileChooserParams?.mode,
+            "acceptTypes" to fileChooserParams?.acceptTypes?.toList(),
+            "filenameHint" to fileChooserParams?.filenameHint,
+            "isCaptureEnabled" to fileChooserParams?.isCaptureEnabled,
+        )
+        FlWebViewPlugin.invokeMethod(
+            channel, handler, "onShowFileChooser", params
+        ) { result ->
+            if (result is ArrayList<*>) {
+                val list = result.map { v -> Uri.fromFile(File(v as String)) }
+                filePathCallback?.onReceiveValue(list.toTypedArray())
+            }
+        }
+        return true
+    }
+
+    override fun onPermissionRequest(request: PermissionRequest?) {
+        super.onPermissionRequest(request)
+        FlWebViewPlugin.invokeMethod(
+            channel, handler, "onPermissionRequest", request?.resources?.toList()
+        ) { result ->
+            if (result is Boolean && result) {
+                request?.grant(request.resources)
+            } else {
+                request?.deny()
+            }
+
+        }
+    }
+
+    override fun onPermissionRequestCanceled(request: PermissionRequest?) {
+        super.onPermissionRequestCanceled(request)
+        FlWebViewPlugin.invokeMethod(
+            channel, handler, "onPermissionRequestCanceled", request?.resources?.toList()
+        )
     }
 
 }