+++ lib/fl_webview.dart
@@ -2,6 +2,5 @@ library fl_webview;
 
 export 'src/extended_web_view.dart';
 export 'src/web_view.dart';
-export 'src/method_channel.dart';
 export 'src/params.dart';
 export 'src/platform_view.dart';